refactoring into static class done
