+++ src/squlearn/expectation_operator/expectation_operator_base.py
@@ -356,11 +356,11 @@ class ExpectationOperatorBase(ABC):
                 """
                 if self._op1 == self._op2:
                     paulis_op = self._op1.get_pauli(parameters)
-                    return OpTree.derivatives.simplify(paulis_op + paulis_op)
+                    return OpTree.simplify(paulis_op + paulis_op)
                 else:
                     paulis_op1 = self._op1.get_pauli(parameters[: self._op1.num_parameters])
                     paulis_op2 = self._op2.get_pauli(parameters[self._op1.num_parameters :])
-                    return OpTree.derivatives.simplify(paulis_op1 + paulis_op2)
+                    return OpTree.simplify(paulis_op1 + paulis_op2)
 
         return AddedExpectationOperator(self, x)
 
@@ -475,10 +475,10 @@ class ExpectationOperatorBase(ABC):
                 """
                 if self._op1 == self._op2:
                     paulis_op = self._op1.get_pauli(parameters)
-                    return optree_simplify(paulis_op.compose(paulis_op))
+                    return OpTree.simplify(paulis_op.compose(paulis_op))
                 else:
                     paulis_op1 = self._op1.get_pauli(parameters[: self._op1.num_parameters])
                     paulis_op2 = self._op2.get_pauli(parameters[self._op1.num_parameters :])
-                    return optree_simplify(paulis_op1.compose(paulis_op2))
+                    return OpTree.simplify(paulis_op1.compose(paulis_op2))
 
         return MultipliedExpectationOperator(self, x)
