+++ src/squlearn/expectation_operator/expectation_operator_derivatives.py
@@ -14,9 +14,6 @@ from ..util.optree.optree import (
     OpTreeLeafOperator,
     OpTree,
 )
-#from ..util.optree.optree_derivative import optree_simplify, optree_derivative
-from ..util.optree.optree_evaluate import optree_assign_parameters
-
 
 class ExpectationOperatorDerivatives:
     r"""Class for calculating derivatives of expectation operators.
@@ -255,7 +252,7 @@ class ExpectationOperatorDerivatives:
                 else:
                     raise ValueError("Unknown type in recursive_squaring:", type(op))
 
-            O2 = OpTree.derivatives.simplify(recursive_squaring(self._optree_start))
+            O2 = OpTree.simplify(recursive_squaring(self._optree_start))
 
             # If caching is enabled, store in the dictionary
             if self._optree_caching == True:
@@ -294,7 +291,7 @@ class ExpectationOperatorDerivatives:
         return_list = []
         for p in param_op_inp:
             dic = dict(zip(self._parameter_vector, p))
-            return_list.append(optree_assign_parameters(operator, dic))
+            return_list.append(OpTree.assign_parameters(operator, dic))
 
         if multi_param_op:
             return OpTreeNodeList(return_list)
@@ -324,7 +321,7 @@ def operator_differentiation(
 
     if len(parameters) == 1:
         # In case of a single parameter no array has to be returned
-        return OpTree.derivatives.simplify(OpTree.derivatives.derivative(optree, parameters).children[0])
+        return OpTree.simplify(OpTree.derivatives.derivative(optree, parameters).children[0])
     else:
         # Check if the same variables are the same type
         params_name = parameters[0].name.split("[", 1)[0]
@@ -332,4 +329,4 @@ def operator_differentiation(
             if p.name.split("[", 1)[0] != params_name:
                 raise TypeError("Differentiable variables are not the same type.")
 
-        return OpTree.derivatives.simplify(OpTree.derivatives.derivative(optree, parameters))
+        return OpTree.simplify(OpTree.derivatives.derivative(optree, parameters))
