+++ src/squlearn/qnn/loss.py
@@ -346,7 +346,7 @@ class SquaredLoss(LossBase):
         and ground_truth as
 
         .. math::
-            2 * \sum_i w_i \left|f\left(x_i\right)-f_ref\left(x_i\right)\right|
+           \sum_j \sum_i w_i \left(f\left(x_i\right)-f_ref\left(x_i\right)\right) \frac{\partial f(x_i)}{\partial p_j}
 
         Args:
             value_dict (dict): Contains calculated values of the model
