+++ src/squlearn/util/optree/__init__.py
@@ -9,15 +9,6 @@ from .optree import (
     OpTree,
 )
 
-from .optree_evaluate import (
-    evaluate_sampler,
-    evaluate_estimator,
-    evaluate_expectation_tree_from_estimator,
-    evaluate_expectation_tree_from_sampler,
-    transform_to_zbasis,
-    optree_assign_parameters,
-)
-
 
 __all__ = [
     "OpTreeNodeList",
@@ -27,11 +18,5 @@ __all__ = [
     "OpTreeLeafContainer",
     "OpTreeLeafExpectationValue",
     "OpTreeLeafMeasuredOperator",
-    "OpTree",
-    "evaluate_sampler",
-    "evaluate_estimator",
-    "evaluate_expectation_tree_from_estimator",
-    "evaluate_expectation_tree_from_sampler",
-    "transform_to_zbasis",
-    "optree_assign_parameters",
+    "OpTree"
 ]
